requirements.txt
uuid~=1.30
Flask~=3.1.0
python-dotenv~=1.0.1
requests~=2.32.3